elf: increase resolution of Atom.input_section_index to u32

+++ src/link/Elf/Atom.zig
@@ -14,7 +14,7 @@ size: u64 = 0,
 alignment: Alignment = .@"1",
 
 /// Index of the input section.
-input_section_index: u16 = 0,
+input_section_index: u32 = 0,
 
 /// Index of the output section.
 output_section_index: u16 = 0,
